feat: 게시글 가중치 설정

+++ src/main/java/com/kakaotech/ott/ott/post/infrastructure/entity/PostEntity.java
@@ -57,7 +57,7 @@ public class PostEntity extends AuditEntity {
     private int scrapCount;
 
     @Column(name = "weight", nullable = false)
-    private int weight;
+    private double weight;
 
     public Post toDomain() {
         return Post.builder()
